+++ src/server.ts
@@ -1,15 +1,16 @@
-import { Device } from './Device';
-import { Discovery } from './Discovery';
+import { Device, Discovery } from './';
 
 const discovery = new Discovery();
 
 discovery.onHello((device: Device) => {
-    console.log(`${new Date().toLocaleTimeString()} - Hello from ${device.serialNumber} on ${device.address}:${device.port}`);
-    console.log(`JSON:\n${JSON.stringify(device)}`);
+    console.log(`${new Date().toLocaleTimeString()} - Hello from ${device.serialNumber} on ${device.address}`);
+    console.log(`\tport: ${device.port}`);
+    console.log(`\tmodel: ${device.modelName}`);
 });
 
 discovery.onGoodbye((device: Device) => {
     console.log(`${new Date().toLocaleTimeString()} - Goodbye from ${device.serialNumber} on ${device.address}`);
 });
 
-discovery.start();
+discovery.start()
+    .then(() => discovery.search());