+++ test/root-description/RootDescription.spec.ts
@@ -7,9 +7,9 @@ chai.should();
 const should = chai.should();
 
 describe('when parsing root description', () => {
-    it('should return remote address', () => {
+    it('should return remote address', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
@@ -19,12 +19,12 @@ describe('when parsing root description', () => {
 
     it('should return friendly name', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getFriendlyNameAsync();
+        const actual = subject.friendlyName;
 
         // Assert
         actual.should.equal('AXIS M1014 - ACCC8E270AD8');
@@ -32,12 +32,12 @@ describe('when parsing root description', () => {
 
     it('should return model description', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getModelDescriptionAsync();
+        const actual = subject.modelDescription;
 
         // Assert
         (actual as string).should.equal('AXIS M1014 Fixed Network Camera');
@@ -45,12 +45,12 @@ describe('when parsing root description', () => {
 
     it('should not return model description if missing', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_REQUIRED_PROPERTIES);
 
         // Act
-        const actual = await subject.getModelDescriptionAsync();
+        const actual = subject.modelDescription;
 
         // Assert
         should.not.exist(actual);
@@ -58,12 +58,12 @@ describe('when parsing root description', () => {
 
     it('should return model name', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getModelNameAsync();
+        const actual = subject.modelName;
 
         // Assert
         actual.should.equal('AXIS M1014');
@@ -71,12 +71,12 @@ describe('when parsing root description', () => {
 
     it('should return model number', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getModelNumberAsync();
+        const actual = subject.modelNumber;
 
         // Assert
         (actual as string).should.equal('M1014');
@@ -84,12 +84,12 @@ describe('when parsing root description', () => {
 
     it('should not return model number if missing', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_REQUIRED_PROPERTIES);
 
         // Act
-        const actual = await subject.getModelNumberAsync();
+        const actual = subject.modelNumber;
 
         // Assert
         should.not.exist(actual);
@@ -97,12 +97,12 @@ describe('when parsing root description', () => {
 
     it('should return serial number', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getSerialNumberAsync();
+        const actual = subject.serialNumber;
 
         // Assert
         (actual as string).should.equal('ACCC8E270AD8');
@@ -110,12 +110,12 @@ describe('when parsing root description', () => {
 
     it('should not return serial number if missing', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_REQUIRED_PROPERTIES);
 
         // Act
-        const actual = await subject.getSerialNumberAsync();
+        const actual = subject.serialNumber;
 
         // Assert
         should.not.exist(actual);
@@ -123,12 +123,12 @@ describe('when parsing root description', () => {
 
     it('should return presentation URL', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_DEFAULT_HTTP_PORT);
 
         // Act
-        const actual = await subject.getPresentationUrlAsync();
+        const actual = subject.presentationUrl;
 
         // Assert
         (actual as string).should.equal('http://192.168.1.102:80/');
@@ -136,12 +136,12 @@ describe('when parsing root description', () => {
 
     it('should not return presentation URL if missing', async () => {
         // Arrange
-        const subject = new RootDescription(
+        const subject = await RootDescription.parse(
             ObjectMother.REMOTE_ADDRESS,
             ObjectMother.ROOT_DESCRIPTION_REQUIRED_PROPERTIES);
 
         // Act
-        const actual = await subject.getPresentationUrlAsync();
+        const actual = subject.presentationUrl;
 
         // Assert
         should.not.exist(actual);