+++ CHANGELOG.md
@@ -5,17 +5,28 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- [#2](https://github.com/FantasticFiasco/axis-discovery-ssdp/issues/2) - Calling `Discovery.stop()` stops listening for SSDP advertisements
+
+### Changed
+
+- Replaced `null` with `undefined` according to [TypeScript guidelines](https://github.com/Microsoft/TypeScript/wiki/Coding-guidelines#null-and-undefined)
+
 ## 1.0.2 - 2017-01-26
 
 ### Fixed
-- [#40](https://github.com/FantasticFiasco/axis-discovery-ssdp/issues/40) - Fixed issue where calling `Discovery.search` didn't trigger a new search
+
+- [#40](https://github.com/FantasticFiasco/axis-discovery-ssdp/issues/40) - Calling `Discovery.search()` didn't trigger a new search
 
 ## 1.0.1 - 2016-12-06
 
 ### Fixed
+
 - Updated `README.md` in package
 
 ## 1.0.0 - 2016-12-04
 
 ### Added
+
 - Support for discovering [Axis Communications](http://www.axis.com/) devices on the network using SSDP